Simplified the JSON handling.
As per issue raised by @andimik there may be issues with unicode handling of station names with extended character sets.
This causes the nlohmann::json library to raise an exception. I've changed this so it will dump the name as is which may
be correct, may be not. But it shouldn't crash and the JSON should be valid.

Forcing Österreich into the station list does not cause it to crash and the JSON is properly formed.

The fib-processor.cpp code uses the function "toStringUsingCharset" to convert from the
characters and character set broadcast to unicode (basically a standard C++ string).

Issues regarding encoding of names was discussed here in 2017:
https://github.com/JvanKatwijk/dab-cmdline/issues/27